Move files to reorganise code structure
