+++ src/main/java/seedu/bookmark/model/wordstore/WordComparator.java
@@ -1,4 +1,4 @@
-package seedu.bookmark.algo.word;
+package seedu.bookmark.model.wordstore;
 
 import java.util.Comparator;
