--- src/test/java/seedu/bookmark/algo/wordbank/GenreWordStoreTest.java
@@ -1,4 +1,0 @@
-package seedu.bookmark.algo.wordbank;
-
-public class GenreWordStoreTest {
-}
