--- src/test/java/seedu/bookmark/algo/wordbank/TagWordStoreTest.java
@@ -1,4 +1,0 @@
-package seedu.bookmark.algo.wordbank;
-
-public class TagWordStoreTest {
-}
