+++ src/test/java/seedu/bookmark/logic/commands/AddCommandTest.java
@@ -14,7 +14,7 @@ import java.util.function.Predicate;
 import org.junit.jupiter.api.Test;
 
 import javafx.collections.ObservableList;
-import seedu.bookmark.algo.WordBank;
+import seedu.bookmark.model.WordBank;
 import seedu.bookmark.commons.core.GuiSettings;
 import seedu.bookmark.logic.commands.exceptions.CommandException;
 import seedu.bookmark.model.Library;
