+++ src/test/java/seedu/bookmark/model/WordBankTest.java
@@ -1,4 +1,4 @@
-package seedu.bookmark.algo;
+package seedu.bookmark.model;
 
 public class WordBankTest {
