+++ src/test/java/seedu/bookmark/model/wordstore/TagWordStoreTest.java
@@ -1,0 +1,4 @@
+package seedu.bookmark.model.wordstore;
+
+public class TagWordStoreTest {
+}
